feat(presentation/error): move error component to your folder
