+++ src/presentation/components/index.ts
@@ -1,1 +1,2 @@
 export { default as ListCard } from './listCard';
+export { default as Error } from './error';
